+++ src/test/java/com/datumbox/applications/nlp/TextClassifierTest.java
@@ -118,9 +118,11 @@ public class TextClassifierTest extends BaseTest {
         }
         
         List<Object> expResult = Arrays.asList("negative","positive");
+        int i = 0;
         for(Integer rId : validationDataset.index()) {
             Record r = validationDataset.get(rId);
-            assertEquals(expResult.get(rId), r.getYPredicted());
+            assertEquals(expResult.get(i), r.getYPredicted());
+            ++i;
         }
         
         instance.delete();